Fetch latest tweet from twitter API
I do not want to post a tweet if it is the exact same as the last tweet
I posted.

This script fetches the latest tweet on the @fullyjabbed account to
compare with the newly generated tweet.

It's untested because I want the duplicates to stop ASAP and I dont
intend to add any error handling. I'll just let the github action fail
if something goes wrong.