Make workflows easier to reuse
Some folks have asked for bots for other countries. I'd like it to be
easier to set up a new bot, for either me or a rando forking this.

So I've moved all the non-secrets that change between bots to
environment variables declared at the top of the job.

I've kept the twitter auth stuff where it is because I dont want secrets
passed to steps that dont need them.